Crate solana_tls_utils

Collection of TLS related code fragments that end up popping up everywhere where quic is used. Aggregated here to avoid bugs due to conflicting implementations of the same functionality.

Structs

QuicClientCertificate

SkipClientVerification

Implementation of ClientCertVerifier that ignores the server certificate. Yet still checks the TLS signatures.

SkipServerVerification

Implementation of ServerCertVerifier that ignores the server certificate. Yet still checks the TLS signatures. This is useful for turbine (where server verification is not feasible) and for tests Logic mostly copied from rustls examples.

Functions

crypto_provider

get_pubkey_from_tls_certificate

new_dummy_x509_certificate

socket_addr_to_quic_server_name

Translate a SocketAddr into a valid SNI for the purposes of QUIC connection

tls_client_config_builder

tls_server_config_builder